requirements.txt
absl-py==0.6.1
astor==0.7.1
astroid==2.3.3
backcall==0.1.0
beautifulsoup4==4.8.1
bleach==3.0.2
blis==0.4.1
boto3==1.10.42
botocore==1.13.42
Bottleneck==1.3.1
catalogue==0.0.8
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cycler==0.10.0
cymem==2.0.3
dataclasses==0.7
decorator==4.3.0
defusedxml==0.5.0
docutils==0.15.2
entrypoints==0.2.3
fastai==1.0.59
fastprogress==0.1.22
funcy==1.11
gast==0.2.0
grpcio==1.16.1
h5py==2.8.0
idna==2.8
imbalanced-learn==0.4.3
importlib-metadata==1.3.0
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.21
jedi==0.13.2
Jinja2==2.10
jmespath==0.9.4
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
lazy-object-proxy==1.4.3
Markdown==3.0.1
MarkupSafe==1.1.0
matplotlib==3.0.2
mccabe==0.6.1
metaflow==2.0.1
mistune==0.8.4
more-itertools==8.0.2
murmurhash==1.0.2
nbconvert==5.4.0
nbformat==4.4.0
nltk==3.4
notebook==5.7.4
numexpr==2.7.0
numpy==1.15.4
nvidia-ml-py3==7.352.0
packaging==19.2
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==6.2.1
plac==1.1.3
preshed==3.0.2
progressbar2==3.39.2
prometheus-client==0.5.0
prompt-toolkit==2.0.7
protobuf==3.6.1
PTable==0.9.2
ptyprocess==0.6.0
Pygments==2.3.1
pylint==2.4.4
pyparsing==2.3.0
python-dateutil==2.7.5
python-utils==2.3.0
pytz==2018.7
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.3
requests==2.22.0
s3transfer==0.2.1
scikit-learn==0.20.2
scipy==1.1.0
Send2Trash==1.5.0
singledispatch==3.4.0.3
six==1.11.0
soupsieve==1.9.5
spacy==2.2.3
srsly==0.2.0
tensorboard==1.12.0
tensorflow==1.12.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
textblob==0.15.2
thinc==7.3.1
torch==1.3.1
torchvision==0.4.2
tornado==5.1.1
tqdm==4.31.1
traitlets==4.3.2
typed-ast==1.4.0
urllib3==1.25.7
wasabi==0.4.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
wrapt==1.11.2
xgboost==0.81
zipp==0.6.0
